+++ src/Generators/PackageGenerator.php
@@ -3,12 +3,39 @@
 namespace Zaynasheff\DocumentGenerator\Generators;
 
 use Zaynasheff\DocumentGenerator\DocumentPackage;
+use Zaynasheff\DocumentGenerator\Factories\DocumentGeneratorFactory;
+use Zaynasheff\DocumentGenerator\Package\Document;
 use Zaynasheff\DocumentGenerator\Result\PackageResult;
 
-class PackageGenerator
+final class PackageGenerator
 {
-    public function generate(DocumentPackage $package): PackageResult
-    {
-        return new PackageResult;
+    public function generate(
+        DocumentPackage $package
+    ): PackageResult {
+
+        $result = new PackageResult;
+
+        $factory = new DocumentGeneratorFactory;
+
+        foreach ($package->items() as $item) {
+
+            if (! $item instanceof Document) {
+                continue;
+            }
+
+            $generator = $factory->make($item);
+
+            $generator->output(
+                $package->outputDirectory()
+            );
+
+            $generationResult = $generator->generate();
+
+            $result->addResult(
+                $generationResult
+            );
+        }
+
+        return $result;
     }
 }